+++ CarsDbApp.Logic/ManagerLogic.cs
@@ -18,16 +18,38 @@ namespace DYRQO6_HFT_2022231.Logic
         }
         public void Create(Manager item)
         {
+            if (item.Name == null)
+            {
+                throw new NullReferenceException("Manager's name cannot be null!");
+            }
+            if (item.Name.Length > 150)
+            {
+                throw new ArgumentException("Please use less characters!(max:150)");
+            }
+            if (item.Age > 90 || item.Age < 18)
+            {
+                throw new ArgumentException("Minimum age is 18, Max age is 90. Please stay inside the scope!");
+            }
             manrepo.Create(item);
         }
 
         public void Delete(int id)
         {
+            var manager = this.manrepo.Read(id);
+            if (manager == null)
+            {
+                throw new ArgumentException("Manager does not exist!");
+            }
             manrepo.Delete(id);
         }
 
         public Manager Read(int id)
         {
+            var manager = this.manrepo.Read(id);
+            if (manager == null)
+            {
+                throw new ArgumentException("Manager does not exist!");
+            }
             return manrepo.Read(id);
         }
 
@@ -38,6 +60,18 @@ namespace DYRQO6_HFT_2022231.Logic
 
         public void Update(Manager item)
         {
+            if (item.Name == null)
+            {
+                throw new NullReferenceException("Manager's name cannot be null");
+            }
+            if (item.Name.Length > 150)
+            {
+                throw new ArgumentException("Please use less characters!(max:150)");
+            }
+            if (item.Age > 90 || item.Age < 18)
+            {
+                throw new ArgumentException("Minimum age is 18, Max age is 90. Please stay inside the scope!");
+            }
             manrepo.Update(item);
         }
     }
